Refactor feed/ and users/ to separate microservices and separate containers
* Created `users-microservice` directory and places the source code
files that handles users/ endpoints. Added Dockerfile to the same
directory to handle creating the users microservice container.

* Created `feed-microservice` directory and places the source code files
 that handles feed/ endpoints. Added Dockerfile to the same directory to
  handle creating the feed microservice container.

* Added two container section in docker-compose.yml for building and
launching the two containers for each microservice to run with rest of
the services when we launch our stack.

* Configured nginx reverse proxy to redirect the api requests to the
correct microservice in order to work with users/ and feed/ endpoints
from the frontend.